v2, refactors all helper classes in spec to mutant_spec_helpers.rb'
